+++ docs/lib/snippets/drift_files/getting_started/database.dart
@@ -1,0 +1,32 @@
+import 'dart:io';
+
+import 'package:drift/drift.dart';
+// These imports are used to open the database
+import 'package:drift/native.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:path/path.dart' as p;
+
+part 'database.g.dart';
+
+@DriftDatabase(
+  // relative import for the drift file. Drift also supports `package:`
+  // imports
+  include: {'tables.drift'},
+)
+class AppDb extends _$AppDb {
+  AppDb() : super(_openConnection());
+
+  @override
+  int get schemaVersion => 1;
+}
+
+LazyDatabase _openConnection() {
+  // the LazyDatabase util lets us find the right location for the file async.
+  return LazyDatabase(() async {
+    // put the database file, called db.sqlite here, into the documents folder
+    // for your app.
+    final dbFolder = await getApplicationDocumentsDirectory();
+    final file = File(p.join(dbFolder.path, 'db.sqlite'));
+    return NativeDatabase(file);
+  });
+}