+++ drift/lib/src/runtime/query_builder/schema/column_impl.dart
@@ -316,10 +316,7 @@ class GeneratedColumnWithTypeConverter<D, S extends Object>
           hasAutoIncrement: hasAutoIncrement,
         );
 
-  /// Compares this column against the mapped [dartValue].
-  ///
-  /// The value will be mapped using the [converter] applied to this column.
-  Expression<bool> equalsValue(D? dartValue) {
+  S? _mapDartValue(D? dartValue) {
     S? mappedValue;
 
     if ($nullable) {
@@ -340,7 +337,56 @@ class GeneratedColumnWithTypeConverter<D, S extends Object>
           "This non-nullable column can't be equal to `null`.", 'dartValue');
     }
 
-    return mappedValue == null ? this.isNull() : equals(mappedValue);
+    return mappedValue;
+  }
+
+  /// Compares this column against the mapped [dartValue].
+  ///
+  /// The value will be mapped using the [converter] applied to this column.
+  /// Unlike [Expression.equals], this handles nullability with semantics one
+  /// might expect in Dart: `null` is equal to `null`.
+  Expression<bool> equalsValue(D? dartValue) {
+    final mappedValue = _mapDartValue(dartValue);
+    if (mappedValue == null) {
+      return this.isNull();
+    } else {
+      return this.isNotNull() & equals(mappedValue);
+    }
+  }
+
+  /// An expression that is true if `this` resolves to any of the values in
+  /// [values].
+  ///
+  /// The values will be mapped using the [converter] applied to this column.
+  /// Unlike [Expression.isIn], this method will also handle nullability with
+  /// semantics on might expect in Dart. If [values] contains `null` and this
+  /// column is nullable, [isInValues] evaluates to `true`.
+  Expression<bool> isInValues(Iterable<D> values) {
+    final mappedValues = values.map(_mapDartValue);
+    final result = isIn(mappedValues.whereNotNull());
+
+    final hasNulls = mappedValues.any((e) => e == null);
+    if (hasNulls) {
+      return result | this.isNull();
+    } else {
+      return result & this.isNotNull();
+    }
+  }
+
+  /// An expression that is true if `this` does not resolve to any of the values
+  /// in [values].
+  ///
+  /// The values will be mapped using the [converter] applied to this column.
+  Expression<bool> isNotInValues(Iterable<D> values) {
+    final mappedValues = values.map(_mapDartValue);
+    final result = isNotIn(mappedValues.whereNotNull());
+
+    final hasNulls = mappedValues.any((e) => e == null);
+    if (hasNulls) {
+      return result & this.isNotNull();
+    } else {
+      return result | this.isNull();
+    }
   }
 }